Migrate llvm (#761)
* first draft


* first draft


* lambda with KrnlBuilder


* support completed for Gemm


* final edits


* remove comments


* remove most krnl edsc


* remove krnl in all put RNNs


* start removing in GRU


* removed in GRU


* removed in RNN & LSTM


* remove all krnl uses but their def


* format


* removed all support for krnl EDSC


* initial


* remove EDSC specific to ONNX dialect


* remove std_ and memref_ calls


* added GRU, unfortunately had to update the mlir lit tests as the number of constant generated changed


* partial update for LSTM


* lstm converted and tested


* remove ScopedContext, not all of them yet


* infrastructure used on memcopy


* some matmul conversion


* completed matmul conversion


* finally edsc free


* format


* rename to mathbuilder due to incompatibility & missing fixes in MLIR


* renaming


* first version succeeded in compiling


* fixed lit test ordering issue


* edit versions of LLVM


* remove eliminated code


* remove temp file and renamed it


* updated to July 9th MLIR


* typo in CMakeLists


* responce to comments


* needed lib
