+++ README.md
@@ -109,7 +109,7 @@ Firstly, install MLIR (as a part of LLVM-Project):
 ``` bash
 git clone https://github.com/llvm/llvm-project.git
 # Check out a specific branch that is known to work with ONNX MLIR.
-cd llvm-project && git checkout ebe408ad8003c946ef871b955ab18e64e82697cb && cd ..
+cd llvm-project && git checkout 23dd750279c9e32ea631cc9e92c4413c7a3df60a && cd ..
 ```
 
 [same-as-file]: <> (utils/build-mlir.sh)
@@ -212,7 +212,7 @@ Install MLIR (as a part of LLVM-Project):
 ```shell
 git clone https://github.com/llvm/llvm-project.git
 # Check out a specific branch that is known to work with ONNX MLIR.
-cd llvm-project && git checkout ebe408ad8003c946ef871b955ab18e64e82697cb && cd ..
+cd llvm-project && git checkout 23dd750279c9e32ea631cc9e92c4413c7a3df60a && cd ..
 ```
 
 [same-as-file]: <> (utils/build-mlir.cmd)
