+++ src/Conversion/KrnlToLLVM/CMakeLists.txt
@@ -6,9 +6,11 @@ add_onnx_mlir_library(OMKrnlToLLVM
   LINK_LIBS PUBLIC
   OMSupport
   MLIRAffineToStandard
+  MLIRMemRefToLLVM
   MLIRSCFToStandard
   MLIRShapeToStandard
   MLIRStandardOpsTransforms
+  MLIRStandardToLLVM
   MLIRVectorToLLVM
   onnx
   )
