+++ src/Conversion/ONNXToKrnl/Tensor/Tile.cpp
@@ -99,7 +99,7 @@ struct ONNXTileOpLowering : public ConversionPattern {
 
     for (int i = 0; i < outputRank; i++) {
       // Scope is created for each dimension because they are independent
-      IndexExprScope IEScope(&rewriter, loc);
+      IndexExprScope IEScope(rewriter, loc);
       DimIndexExpr index(outputLoops.getInductionVar(i));
       MemRefBoundsIndexCapture inputBounds(input);
       DimIndexExpr dimSize(inputBounds.getDim(i));
