+++ src/Dialect/Krnl/KrnlOps.hpp
@@ -25,12 +25,10 @@
 #include "src/Interface/SpecializedKernelOpInterface.hpp"
 #include "llvm/ADT/TypeSwitch.h"
 
-// TODO rename to MLIR file
-#include "src/Dialect/ONNX/IndexExpr.hpp"
-#include "src/Dialect/ONNX/TmpMlirUtils.hpp"
-
 #include "src/Dialect/Krnl/KrnlHelper.hpp"
 #include "src/Dialect/Krnl/KrnlTypes.hpp"
+#include "src/Dialect/ONNX/IndexExpr.hpp"
+#include "src/Dialect/ONNX/MLIRDialectBuilder.hpp"
 
 namespace mlir {
 class KrnlOpsDialect : public Dialect {
