+++ test/mlir/onnx/onnx_bundle_memory_pool.mlir
@@ -10,14 +10,14 @@ func @test_bundle_memory_pool(%arg0: tensor<10x10xf32>, %arg1: tensor<10x20xf32>
   return %5 : tensor<10x20xf32>
 
   // CHECK-LABEL: test_bundle_memory_pool
-  // CHECK: [[CONST0:%.+]] = constant 0 : i64
-  // CHECK: [[CONST00:%.+]] = constant 0.000000e+00 : f32
-  // CHECK: [[CONST2400:%.+]] = constant 2400 : i64
-  // CHECK: [[CONST2000:%.+]] = constant 2000 : i64
-  // CHECK: [[CONST1200:%.+]] = constant 1200 : i64
-  // CHECK: [[CONST400:%.+]] = constant 400 : i64
-  // CHECK: [[RES:%.+]] = memref.alloc() : memref<10x20xf32>
-  // CHECK: [[MEMPOOL:%.+]] = memref.alloc() : memref<3200xi8>
+  // CHECK-DAG: [[CONST0:%.+]] = constant 0 : i64
+  // CHECK-DAG: [[CONST00:%.+]] = constant 0.000000e+00 : f32
+  // CHECK-DAG: [[CONST2400:%.+]] = constant 2400 : i64
+  // CHECK-DAG: [[CONST2000:%.+]] = constant 2000 : i64
+  // CHECK-DAG: [[CONST1200:%.+]] = constant 1200 : i64
+  // CHECK-DAG: [[CONST400:%.+]] = constant 400 : i64
+  // CHECK-DAG: [[RES:%.+]] = memref.alloc() : memref<10x20xf32>
+  // CHECK-DAG: [[MEMPOOL:%.+]] = memref.alloc() : memref<3200xi8>
   // CHECK: "krnl.getref"([[MEMPOOL]], [[CONST2400]]) : (memref<3200xi8>, i64) -> memref<10x20xf32>
   // CHECK: "krnl.getref"([[MEMPOOL]], [[CONST2000]]) : (memref<3200xi8>, i64) -> memref<10x10xf32>
   // CHECK: "krnl.getref"([[MEMPOOL]], [[CONST1200]]) : (memref<3200xi8>, i64) -> memref<10x20xf32>
